backend/requirements.txt
flask
peewee
PyMySQL
python-dotenv
flask-cors
